requirements.txt
streamlit
google-generativeai
ddgs
python-dotenv
pandas
fastapi
uvicorn
httpx
redis
pydantic
pydantic-settings
lxml
groq
apscheduler
streamlit-authenticator
feedparser
newsapi-python
ruff
pytest
pytest-asyncio
httpx
